# sys-service/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Aevons 系统服务接口文档，提供用户、角色、部门、岗位、菜单、字典、语言、配置、通知公告和终端应用等系统管理能力",
        "title": "Aevons System Service API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:10702",
    "basePath": "/api/v1/sys",
    "paths": {
...
        "/api/v1/sys/dict": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "新增字典类型",
                "parameters": [
                    {
                        "description": "新增字典类型",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDictDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sys/dict/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "查询可用字典类型列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/dict/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "分页查询字典类型",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sys/dict/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "获取字典类型详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "修改字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改字典类型",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateDictDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "删除字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/lang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "查询可用语言列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "新增语言",
                "parameters": [
                    {
                        "description": "新增语言",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLangDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/lang/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "查询语言列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/lang/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "分页查询语言",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sys/lang/{ids}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "批量删除语言",
                "parameters": [
                    {
                        "type": "string",
                        "description": "逗号分隔的语言ID",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/lang/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "获取语言详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "语言ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "修改语言",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "语言ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改语言",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateLangDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateDictDTO": {
            "type": "object",
            "properties": {
                "dictName": {
                    "description": "字典名称",
                    "type": "string",
                    "maxLength": 50
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string",
                    "maxLength": 32
                },
                "isSys": {
                    "description": "系统内置（0否 1是）",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 500
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateLangDTO": {
            "type": "object",
            "required": [
                "isDefault",
                "langCode",
                "langName",
                "sort",
                "status"
            ],
            "properties": {
                "isDefault": {
                    "description": "是否默认语言（0否，1是）",
                    "type": "integer"
                },
                "langCode": {
                    "description": "语言编码（如zh-CN、en-US）",
                    "type": "string",
                    "maxLength": 10
                },
                "langName": {
                    "description": "语言名称（如简体中文、English）",
                    "type": "string",
                    "maxLength": 50
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 200
                },
                "sort": {
                    "description": "排序值（升序）",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常,1停用）",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateDictDTO": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "dictName": {
                    "description": "字典名称",
                    "type": "string",
                    "maxLength": 50
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string",
                    "maxLength": 32
                },
                "id": {
                    "description": "字典主键",
                    "type": "string",
                    "example": "0"
                },
                "isSys": {
                    "description": "系统内置（0否 1是）",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 500
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateLangDTO": {
            "type": "object",
            "required": [
                "id",
                "isDefault",
                "langCode",
                "langName",
                "sort",
                "status"
            ],
            "properties": {
                "id": {
                    "description": "编号",
                    "type": "string",
                    "example": "0"
                },
                "isDefault": {
                    "description": "是否默认语言（0否，1是）",
                    "type": "integer"
                },
                "langCode": {
                    "description": "语言编码（如zh-CN、en-US）",
                    "type": "string",
                    "maxLength": 10
                },
                "langName": {
                    "description": "语言名称（如简体中文、English）",
                    "type": "string",
                    "maxLength": 50
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 200
                },
                "sort": {
                    "description": "排序值（升序）",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常,1停用）",
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer 访问令牌，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}